Move function related to auth out of account controller into new auth controller. Update tests accordingly
