+++ src/controllers/account.js
@@ -21,17 +21,6 @@ module.exports = [
     const errorCodes = errors.errorCodes;
     const { isValid, invalid, isValidPassword, generateToken } = utils;
 
-    const getTokenQuery = (id, token) => {
-      return {
-        _id: id,
-        "tokens.token": token,
-      };
-    };
-
-    const tokenFilter = {
-      tokens: 1,
-    };
-
     return {
       get: async (req, res, next) => {
         try {
@@ -138,82 +127,6 @@ module.exports = [
         } catch (error) {
           next(error);
         }
-      },
-
-      login: async (req, res, next) => {
-        try {
-          const { email, password } = req.body;
-
-          if (!email) {
-            const error = new Error(errorCodes.MISSING_EMAIL);
-            error.name = errorCodes.MISSING_EMAIL;
-            throw error;
-          }
-
-          if (!password) {
-            const error = new Error(errorCodes.MISSING_PASSWORD);
-            error.name = errorCodes.MISSING_PASSWORD;
-            throw error;
-          }
-
-          const account = await accountRepository.findByEmail(email);
-
-          if (!account || !(await bcrypt.compare(password, account.password))) {
-            const error = new Error(errorCodes.LOGIN_FAILED);
-            error.name = errorCodes.LOGIN_FAILED;
-            throw error;
-          }
-
-          const token = await generateToken(account._id.toString());
-          await accountRepository.insertToken(account._id, token);
-
-          res.json({
-            token,
-          });
-        } catch (error) {
-          next(error);
-        }
-      },
-
-      logout: async (req, res, next) => {
-        try {
-          const query = getTokenQuery(req.auth.id, req.auth.token);
-
-          await accountRepository.deleteTokens({ query, filter: tokenFilter });
-          res.send();
-        } catch (error) {
-          next(error);
-        }
-      },
-
-      logoutAll: async (req, res, next) => {
-        try {
-          const query = getTokenQuery(req.auth.id, req.auth.token);
-
-          await accountRepository.deleteTokens({
-            query,
-            filter: tokenFilter,
-            removeAll: true,
-          });
-          res.send();
-        } catch (error) {
-          next(error);
-        }
-      },
-
-      checkToken: async (id, token) => {
-        const opts = {
-          query: getTokenQuery(id, token),
-          lean: true,
-        };
-
-        const login = await accountRepository.get(opts);
-
-        if (login) {
-          return true;
-        }
-
-        return false;
       },
     };
   },
